* ctest.h   (assert_equal): Change types to long to make it support both int and long.

+++ ctest.h
@@ -95,7 +95,7 @@ void CTEST_LOG(char *fmt, ...);
 void assert_str(const char* exp, const char*  real, const char* caller, int line);
 #define ASSERT_STR(exp, real) assert_str(exp, real, __FILE__, __LINE__)
 
-void assert_equal(int exp, int real, const char* caller, int line);
+void assert_equal(long exp, long real, const char* caller, int line);
 #define ASSERT_EQUAL(exp, real) assert_equal(exp, real, __FILE__, __LINE__)
 
 void assert_not_equal(int exp, int real, const char* caller, int line);
@@ -224,9 +224,9 @@ void assert_str(const char* exp, const char*  real, const char* caller, int line
     }
 }
 
-void assert_equal(int exp, int real, const char* caller, int line) {
+void assert_equal(long exp, long real, const char* caller, int line) {
     if (exp != real) {
-        CTEST_ERR("%s:%d  expected %d, got %d", caller, line, exp, real);
+        CTEST_ERR("%s:%d  expected %ld, got %ld", caller, line, exp, real);
         longjmp(ctest_err, 1);
     }
 }
